Move guard clause helpers to detekt-rules-style
These are only used in two rules and have limited utility elsewhere
